This is REVISION has a new Main Engine structure which is completely Object Oriented.
It is still unplayable, but shows the Intro screen as demonstration and runs very stable as it seems.

If you want to help us on new features, use this revision, but read carefully our forum first!!!

If you want to fix bugs on other sections, like Sound, AI or game play, please stick with REVISION 427.

I would appreciate if you use the most recent version and adapt it! Once again read the forum!!!!!!

That's very important!

git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@428 4df4b0f3-56ce-47cb-b001-ed939b7d65a6