I fixed up all of the infoscene type things, now all we need as far as these go is to add the sprites in the different ordering infos (which all work now) like the Garg and Yorp that stand there on EP1, the Scrub that moves around the screen on EP2, and Keen standing at the left side of the screen and shooting that weird yellow thing that comes on screen every now and again (without a ray firing), and then we will be all done with all of the infoscene related stuff other than to make it center properly in the 320x240 mode and any other modes we may create.
This and a few other things make me hesitant to declare a release on v0.3.1 till sometime in January, then we would have time to implement a new demo system, or work on adding the one from Keens 4-6 with support for Keen 1-3.  Multiplayer won't return till after we have finished with Keens 4-6 for certain, it could come before that, but only if I get the time to figure out all of the inner workings of the entire game so that I can make the necessary changes.  We can however release what we have as v0.3.0.7 tomorrow (Christmas) so that everyone can see just how cool this one is (on all systems, bundle, link on website) but we need to synchronize our uploading, so email me when you get ready to release or something like that, and I will upload the file so that you can change what you need to with the website and get it up there.  You should also release for the Wiz Pickle.

git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@746 4df4b0f3-56ce-47cb-b001-ed939b7d65a6