Reworked Menus...
Be careful! Only new game and quit are working properly. This is the first part of the commit. YOu can play the the game, but the options are limited in this revision.

The second part will have less, more functions, better code and everything working...

Don't use this one for releases!

git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@956 4df4b0f3-56ce-47cb-b001-ed939b7d65a6