on iphone, we use just the app-bundle-root-dir as the searchpath
It is even important that we don't use the "Resources" directory. For some strange reasons, this results in a lot of trouble. Esp., I was getting this error when trying to use it:
Failed to launch simulated application: Unknown error.

Also the installation on the iPhone-Simulator didn't seem to work.

See this for some reference:
http://stackoverflow.com/questions/1262898/failed-to-launch-simulated-application-unknown-error/1265168#1265168