deleted unused iphone files
I wasn't sure if I really should delete those as they may be good for later references. Anyway, did it for now, I can also just peek into this Git log.

If it points out that I/we need it once and then, we should move it somewhere to build/Xcode/iphone-src-samples or so.