Mac can't enter Fullscreen at the moment.
Plus I just wanted to do an overhaul of this list a little bit.  In any
case, I saw how you're handling the list of resolutions now, and I
don't think it's really very great, we should really try to find a
middle ground that allows us to check to see if each other these will
work on the current display, and allows just regular window sizes when
in windowed mode.