Further work on iPhone port
I might be straying into the area where we'll have to change things
significantly because of SDL 2.0.  Namely, in creating the window.  But
for now, this change allows the launcher to load and input to be
detected.  Although the game crashes when you try to actually load a
game (done by clicking randomly at the screen because no graphics
appear) so there's more graphics based things that it's not liking.
Also, I don't need the iPhone specific source files anymore.